Seed Databases

+++ database/migrations/2016_07_29_205026_delete_column_email_posts_table.php
@@ -12,9 +12,12 @@ class DeleteColumnEmailPostsTable extends Migration
      */
     public function up()
     {
-        Schema::table('posts', function (Blueprint $table) {
-            $table->dropColumn('email');
-        });
+        if (Schema::hasColumn('posts', 'email'))
+        {
+            Schema::table('posts', function (Blueprint $table) {
+                $table->dropColumn('email');
+            });
+        }
     }
 
     /**
